+++ src/sc/branching/commands/common.py
