+++ src/sc/branching_cli.py
@@ -295,33 +295,7 @@ def checkout(name, force, verify):
     """Checkout a support branch."""
     SCBranching.checkout(BranchType.SUPPORT, name, force, verify)
 
-@cli.group()
-def show():
-    """sc show commands."""
-    pass
-
-@show.command()
-def branch():
-    """Show the current status of branching."""
-    SCBranching.show_branch()
-
-@show.command()
-@click.argument("tag")
-def tag(tag):
-    """Show information about a tag."""
-    pass
-
-@show.command()
-@click.argument("group", required=False)
-def group(group):
-    """Show information about a group."""
-    SCBranching.show_group(group)
-
-@show.command()
-def repo_flow_config():
-    """Show git flow config for all projects."""
-    SCBranching.show_repo_flow_config()
-
+@cli.group
 def tag():
     pass
 
@@ -360,6 +334,38 @@ def push(tag):
 def check(tag):
     """Check if a tag exists on all non READ_ONLY repos."""
     SCBranching.tag_check(tag)
+
+@cli.group()
+def show():
+    """sc show commands."""
+    pass
+
+@show.command()
+def branch():
+    """Show the current status of branching."""
+    SCBranching.show_branch()
+
+@show.command()
+def repo_flow_config():
+    """Show git flow config for all projects."""
+    SCBranching.show_repo_flow_config()
+
+@show.command()
+@click.argument("tag")
+def tag(tag):
+    """Show information about a tag."""
+    SCBranching.tag_show(tag)
+
+@show.command()
+def tags():
+    """List tags on the manifest."""
+    SCBranching.tag_list()
+
+@show.command()
+@click.argument("group", required=False)
+def group(group):
+    """List groups or show information about a group."""
+    SCBranching.group_show(group)
 
 if __name__ == '__main__':
     cli()
